Сделал условия для миграций в отложенных скриптах и правки

+++ database/migrations/2020_04_10_1586470060_create_support_scripts_exceptions_table.php
@@ -7,16 +7,31 @@ use Illuminate\Database\Migrations\Migration;
 class CreateSupportScriptsExceptionsTable extends Migration
 {
     /**
+     * @var string
+     */
+    private $table_name;
+
+    /**
+     * CreateSupportScriptsExceptionsTable constructor.
+     */
+    public function __construct()
+    {
+        $this->table_name = app_config('support_bot.scripts.exceptions_table_name');
+    }
+
+    /**
      * Run the migrations.
      *
      * @return void
      */
     public function up()
     {
-        Schema::create(app_config('support_bot.scripts.exceptions_table_name'), function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('exception');
-        });
+        if(!empty($this->table_name)) {
+            Schema::create($this->table_name, function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('exception');
+            });
+        }
     }
 
     /**
@@ -26,7 +41,9 @@ class CreateSupportScriptsExceptionsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists(app_config('support_bot.scripts.exceptions_table_name'));
+        if(!empty($this->table_name)) {
+            Schema::dropIfExists(app_config('support_bot.scripts.exceptions_table_name'));
+        }
     }
 
 }
