+++ database/migrations/2020_04_7_1586232600_create_support_pending_scripts_table.php
@@ -7,19 +7,34 @@ use Illuminate\Database\Migrations\Migration;
 class CreateSupportPendingScriptsTable extends Migration
 {
     /**
+     * @var string
+     */
+    private $table_name;
+
+    /**
+     * CreateSupportPendingScriptsTable constructor.
+     */
+    public function __construct()
+    {
+        $this->table_name = app_config('support_bot.scripts.table_name');
+    }
+
+    /**
      * Run the migrations.
      *
      * @return void
      */
     public function up()
     {
-        Schema::create(app_config('support_bot.scripts.table_name'), function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedBigInteger('search_id');
-            $table->tinyInteger('step')->default(0);
-            $table->tinyInteger('prev_step')->nullable();
-            $table->dateTime('send_message_at');
-        });
+        if(!empty($this->table_name)) {
+            Schema::create($this->table_name, function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedBigInteger('search_id');
+                $table->tinyInteger('step')->default(0);
+                $table->tinyInteger('prev_step')->nullable();
+                $table->dateTime('send_message_at');
+            });
+        }
     }
 
     /**
@@ -29,7 +44,9 @@ class CreateSupportPendingScriptsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists(app_config('support_bot.scripts.table_name'));
+        if(!empty($this->table_name)) {
+            Schema::dropIfExists(app_config('support_bot.scripts.table_name'));
+        }
     }
 
 }
